begin router implementation
