+++ app/_layout.js
@@ -1,0 +1,12 @@
+import { Stack } from "expo-router/stack";
+export default function Layout() {
+	return (
+		<Stack
+			screenOptions={{
+				headerShown: false
+			}}
+		>
+			<Stack.Screen name="login" />
+		</Stack>
+	);
+}
